requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart>=0.0.5
streamlit==1.28.1
streamlit-extras==0.3.5

# PDF Processing
PyPDF2==3.0.1

# LangChain and AI/ML
langchain>=1.0.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-classic>=1.0.0
langchain-text-splitters>=1.0.0
# langchain-core will be installed automatically as a dependency
openai>=1.3.5

# Vector Store
faiss-cpu==1.8.0

# HTTP Client
requests>=2.32.5,<3.0.0

# Environment Variables
python-dotenv==1.0.0

# Python 3.12+ compatibility (provides distutils)
setuptools>=65.0.0

# CORS Support (included in FastAPI but explicit)
pydantic>=2.7.4,<3.0.0

